refactor(Video): Entity에서 DTO 분리
